requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.6
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.7
django-cors-headers==3.8.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
idna==3.2
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
Pillow==8.3.2
psycopg2-binary==2.8.6
PyJWT==2.1.0
python-dotenv==0.19.0
pytz==2021.1
requests==2.26.0
sqlparse==0.4.2
uritemplate==3.0.1
urllib3==1.26.7
